backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
python-multipart
supabase
pydantic
pydantic-settings
Pillow
python-dotenv
python-jose[cryptography]
httpx
